# PERF: Groupby.quantile