```
from asv_bench.benchmarks.groupby import *
self2 = GroupByMethods()
self2.setup('uint', 'quantile', 'direct', 1)

%timeit self2.time_dtype_as_group('uint', 'quantile', 'direct', 1)
1.03 ms ± 46.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master
328 µs ± 17.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
```